Move GetNBZone to libovsdb/util pkg
Helps remove dependency of util pkg on libovsdbops
