+++ go-controller/pkg/network-controller-manager/network_controller_manager.go
@@ -13,6 +13,7 @@ import (
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube"
 	libovsdbops "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdb/ops"
+	libovsdbutil "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdb/util"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/metrics"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/nbdb"
 	nad "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/network-attach-def-controller"
@@ -306,7 +307,7 @@ func (cm *networkControllerManager) Start(ctx context.Context) error {
 	var zone string
 	var err1 error
 	err := wait.PollUntilContextTimeout(ctx, 250*time.Millisecond, maxTimeout, true, func(ctx context.Context) (bool, error) {
-		zone, err1 = util.GetNBZone(cm.nbClient)
+		zone, err1 = libovsdbutil.GetNBZone(cm.nbClient)
 		if err1 != nil {
 			return false, nil
 		}
